Rename legacy URL resolver files and associated tests
