+++ app/api/resources/routes.py
@@ -17,10 +17,12 @@ limitations under the License.
 
 from fastapi import APIRouter
 
-from app.api.resources import resolver_view, rapid_view, url_resolver_view
+from app.api.resources import resolver_view, rapid_view, legacy_url_resolver_view
 
 router = APIRouter()
 
 router.include_router(resolver_view.router, tags=["resolver"], prefix="/id")
 router.include_router(rapid_view.router, tags=["rapid"], prefix="/rapid")
-router.include_router(url_resolver_view.router, tags=["url-resolver"], prefix="/legacy")
+router.include_router(
+    legacy_url_resolver_view.router, tags=["url-resolver"], prefix="/legacy"
+)
