+++ tests/test_legacy_url_resolver.py
@@ -19,7 +19,7 @@ class TestUrlResolver(unittest.TestCase):
         self.mock_url_resolver_api_url = "/legacy"
         self.genome_uuid = "genome_uuid1"
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_species_home_with_json_response(self, mock_species_lookup):
         """Resolve a species home URL to the new Ensembl species page."""
         mock_species_lookup.return_value = self.genome_uuid
@@ -38,7 +38,7 @@ class TestUrlResolver(unittest.TestCase):
         )
         mock_species_lookup.assert_called_once_with("Homo_sapiens")
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_bare_species_path_with_redirect(self, mock_species_lookup):
         """Resolve a bare species path to the new Ensembl species page."""
         mock_species_lookup.return_value = self.genome_uuid
@@ -56,7 +56,7 @@ class TestUrlResolver(unittest.TestCase):
         )
         mock_species_lookup.assert_called_once_with("Crocodylus_porosus")
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_bare_species_path_without_uuid_redirects_to_archive(
         self, mock_species_lookup
     ):
@@ -76,7 +76,7 @@ class TestUrlResolver(unittest.TestCase):
         )
         mock_species_lookup.assert_called_once_with("Homo_sapiens")
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_gene_summary_with_redirect(self, mock_species_lookup):
         """Resolve a supported gene URL to a permanent new Ensembl redirect."""
         mock_species_lookup.return_value = self.genome_uuid
@@ -98,7 +98,7 @@ class TestUrlResolver(unittest.TestCase):
             (f"{ENSEMBL_URL}/entity-viewer/{self.genome_uuid}" "/gene:ENSG00000012048"),
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_gene_summary_with_uuid_lookup_value(self, mock_species_lookup):
         """Resolve URLs when DuckDB returns genome UUID as a UUID object."""
         genome_uuid = UUID("12345678-1234-5678-1234-567812345678")
@@ -126,7 +126,7 @@ class TestUrlResolver(unittest.TestCase):
             },
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_location_view_with_semicolon_query(self, mock_species_lookup):
         """Resolve old-style semicolon-delimited query strings."""
         mock_species_lookup.return_value = self.genome_uuid
@@ -155,7 +155,7 @@ class TestUrlResolver(unittest.TestCase):
             },
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_transcript_summary(self, mock_species_lookup):
         """Resolve a transcript summary URL to the new Ensembl entity viewer."""
         mock_species_lookup.return_value = self.genome_uuid
@@ -183,7 +183,7 @@ class TestUrlResolver(unittest.TestCase):
             },
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_required_query_parameter(self, mock_species_lookup):
         """Return 400 when a URL shape is known but its parameter is missing."""
         mock_species_lookup.return_value = self.genome_uuid
@@ -197,7 +197,7 @@ class TestUrlResolver(unittest.TestCase):
 
         self.assertEqual(response.status_code, 400)
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_unsupported_template(self, mock_species_lookup):
         """Return 404 for spreadsheet rows with no supported new Ensembl mapping."""
         mock_species_lookup.return_value = self.genome_uuid
@@ -217,7 +217,7 @@ class TestUrlResolver(unittest.TestCase):
         self.assertEqual(response.status_code, 404)
         mock_species_lookup.assert_not_called()
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_redirects_to_main_archive(
         self, mock_species_lookup
     ):
@@ -236,7 +236,7 @@ class TestUrlResolver(unittest.TestCase):
             "https://jun2026.archive.ensembl.org/Homo_sapiens/Info/Index",
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_null_species_uuid_redirects_to_main_archive(
         self, mock_species_lookup
     ):
@@ -261,7 +261,7 @@ class TestUrlResolver(unittest.TestCase):
             "?g=ENSG00000012048",
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_redirects_to_plants_archive(
         self, mock_species_lookup
     ):
@@ -282,7 +282,7 @@ class TestUrlResolver(unittest.TestCase):
             "https://eg63-plants.ensembl.org/Arabidopsis_thaliana/Info/Index",
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_redirects_to_metazoa_archive(
         self, mock_species_lookup
     ):
@@ -303,7 +303,7 @@ class TestUrlResolver(unittest.TestCase):
             "https://eg63-metazoa.ensembl.org/Caenorhabditis_elegans/Info/Index",
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_redirects_to_fungi_archive(
         self, mock_species_lookup
     ):
@@ -324,7 +324,7 @@ class TestUrlResolver(unittest.TestCase):
             "https://eg63-fungi.ensembl.org/Saccharomyces_cerevisiae/Info/Index",
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_redirects_to_protists_archive(
         self, mock_species_lookup
     ):
@@ -345,7 +345,7 @@ class TestUrlResolver(unittest.TestCase):
             "https://eg63-protists.ensembl.org/Plasmodium_falciparum/Info/Index",
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_redirects_to_bacteria_archive(
         self, mock_species_lookup
     ):
@@ -372,7 +372,7 @@ class TestUrlResolver(unittest.TestCase):
             ),
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_archive_preserves_query_and_fragment(
         self, mock_species_lookup
     ):
@@ -397,7 +397,7 @@ class TestUrlResolver(unittest.TestCase):
             "?r=1:1-100#content",
         )
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_missing_species_unknown_host_does_not_redirect(
         self, mock_species_lookup
     ):
@@ -414,7 +414,7 @@ class TestUrlResolver(unittest.TestCase):
         self.assertEqual(response.status_code, 404)
         self.assertNotIn("location", response.headers)
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_unknown_page_with_json_response(self, mock_species_lookup):
         """Return JSON error for unsupported one-segment legacy paths."""
         mock_species_lookup.side_effect = SpeciesMappingNotFoundError("not found")
@@ -430,7 +430,7 @@ class TestUrlResolver(unittest.TestCase):
         self.assertNotIn("location", response.headers)
         mock_species_lookup.assert_called_once_with("foo")
 
-    @patch("app.api.resources.url_resolver_view.get_genome_uuid_from_species_url")
+    @patch("app.api.resources.legacy_url_resolver_view.get_genome_uuid_from_species_url")
     def test_resolve_unknown_page_with_html_interstitial(self, mock_species_lookup):
         """Render a choice page for browser users on unsupported legacy paths."""
         mock_species_lookup.side_effect = SpeciesMappingNotFoundError("not found")
